requirements_3.12.txt
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
fonttools==4.53.1
hexdump==3.3
JPype1==1.5.0
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.2
netifaces-plus==0.12.2
numpy==2.1.1
packaging==24.1
pillow==10.4.0
Pygments==2.18.0
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
QtPy==2.4.1
scapy==2.5.0
six==1.16.0
superqt==0.6.7
transitions==0.9.2
typing_extensions==4.12.2
xsdata==24.7
